Convert the code to a "proper" async handler
There are two general approaches for Lambda functions: A non-async handler uses the callback function (or optionally context.fail/context.succeed)
to return the result of the execution, while a async handler simply returns a suitable promise (and potentially uses `async` functions
for implementing that behavior).

The code here was a mix between both styles, which is fine -- but quite confusing. Additionally though this was also using the
(no longer?) documented context.succeed approach, which made it fairly hard to reason about.
Simplify this all by simply using the async style consistently.